+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add command to show report. Closes [#8](https://github.com/ryu1kn/vscode-extension-update-reporter/issues/8) by @jerone in [PR #38](https://github.com/ryu1kn/vscode-extension-update-reporter/pull/38).
 - Add before and after versions. Closes [#4](https://github.com/ryu1kn/vscode-extension-update-reporter/issues/4) by @jerone in [PR #40](https://github.com/ryu1kn/vscode-extension-update-reporter/pull/40).
 
+### Fixed
+- Improve homepage and add repository links. Fixes [#41](https://github.com/ryu1kn/vscode-extension-update-reporter/issues/41) by @jerone in [PR #42](https://github.com/ryu1kn/vscode-extension-update-reporter/pull/42).
+
 ## [1.2.0] - 2022-11-06
 ### Added
 - Marketplace, extension homepage links are always available for easier access to the full information. (thanks to @jerone [PR #29](https://github.com/ryu1kn/vscode-extension-update-reporter/issues/29))